Automatically backed up by Learn

+++ db/migrate/20190712170245_remove_email_add_balance.rb
@@ -1,6 +1,6 @@
 class RemoveEmailAddBalance < ActiveRecord::Migration[5.1]
   def change
     rename_column :users, :email, :balance
-    change_column :users, :balance, :float
+    change_column :users, :balance, :double
   end
 end
